+++ forces_gpu.h
@@ -5,10 +5,10 @@
 // extern "C"{
 // #endif
 
-void launch_forces(const double *rx, const double *ry, const double *rz,
-              double *fx, double *fy, double *fz,
-              double *epot, double *pres, const double *temp,
-              const double rho, const double V, const double L);
+void launch_forces(const float *rx, const float *ry, const float *rz,
+              float *fx, float *fy, float *fz,
+              float *epot, float *pres, const float *temp,
+              const float rho, const float V, const float L);
 
 // #ifdef __cplusplus
 // }
